image_ml/backend/requirements.txt
fastapi
uvicorn
pillow
tflite-runtime
numpy<2.0.0
python-multipart